accurate answers according to the lesson
The lesson introduces the idea that short-circuit evaluations can prevent both expensive calculations and run-time errors. The following question does not propose an answer that is suitable, since it provides both options as a possible answer but only accepts one value. The resolution is to include one or more answer options that contains both correct answers as a compound value.